+++ CHANGELOG.md
@@ -3,6 +3,11 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 #### 1.10.0-SNAPSHOT
 
 - **[UPDATE]** Updated to Dokka 1.5.30.
+- **[FEATURE]** (`ktx-collections`) `isEmpty` and `isNotEmpty` extension methods of LibGDX collections now support
+Kotlin contracts. If they return `true`, the collection is implied not to be null.
+- **[CHANGE]** (`ktx-collections`) As `IdentityMap` now extends `ObjectMap`, some redundant utilities specific to
+`IdentityMap` were removed. From the user point of view, all additional functionalities are covered by the `ObjectMap`
+utilities.
 
 #### 1.10.0-b3